feat: add types for frontend code (#48)
